@@ -10,7 +10,7 @@
 
     <IsPublishable>False</IsPublishable>
 
-    <TargetFrameworks>net7.0-windows10.0.17763.0</TargetFrameworks>
+    <TargetFrameworks>net8.0-windows10.0.17763.0</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>
